versão de deploy no ecs

+++ app/wp-content/plugins/elementor/core/common/modules/connect/apps/base-app.php
@@ -11,7 +11,7 @@ use Elementor\Tracker;
 use Elementor\Utils;
 
 if ( ! defined( 'ABSPATH' ) ) {
-	exit; // Exit if accessed directly
+	exit; // Exit if accessed directly.
 }
 
 abstract class Base_App {
@@ -79,13 +79,16 @@ abstract class Base_App {
 
 		if ( $this->is_connected() ) {
 			$remote_user = $this->get( 'user' );
-			/* translators: %s: Remote user. */
-			$title = sprintf( esc_html__( 'Connected as %s', 'elementor' ), '<strong>' . esc_html( $remote_user->email ) . '</strong>' );
+			$title = sprintf(
+				/* translators: %s: Remote user. */
+				esc_html__( 'Connected as %s', 'elementor' ),
+				'<strong>' . esc_html( $remote_user->email ) . '</strong>'
+			);
 			$label = esc_html__( 'Disconnect', 'elementor' );
 			$url = $this->get_admin_url( 'disconnect' );
 			$attr = '';
 
-			echo sprintf(
+			printf(
 				'%s <a %s href="%s">%s</a>',
 				// PHPCS - the variable $title is already escaped above.
 				$title, // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
@@ -168,7 +171,6 @@ abstract class Base_App {
 
 	public function action_reset() {
 		if ( current_user_can( 'manage_options' ) ) {
-			delete_option( static::OPTION_CONNECT_SITE_KEY );
 			delete_option( 'elementor_remote_info_library' );
 		}
 
@@ -468,7 +470,7 @@ abstract class Base_App {
 			$args
 		);
 
-		if ( is_wp_error( $response ) ) {
+		if ( is_wp_error( $response ) && empty( $options['with_error_data'] ) ) {
 			// PHPCS - the variable $response does not contain a user input value.
 			wp_die( $response, [ 'back_link' => true ] ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 		}
@@ -491,12 +493,20 @@ abstract class Base_App {
 			return new \WP_Error( 422, 'Wrong Server Response' );
 		}
 
+		if ( 201 === $response_code ) {
+			return $body;
+		}
+
 		if ( 200 !== $response_code ) {
 			// In case $as_array = true.
 			$body = (object) $body;
 
 			$message = isset( $body->message ) ? $body->message : wp_remote_retrieve_response_message( $response );
 			$code = (int) ( isset( $body->code ) ? $body->code : $response_code );
+
+			if ( ! $code ) {
+				$code = $response_code;
+			}
 
 			if ( 401 === $code ) {
 				$this->delete();
@@ -506,6 +516,10 @@ abstract class Base_App {
 				if ( $should_retry ) {
 					$this->action_authorize();
 				}
+			}
+
+			if ( isset( $options['with_error_data'] ) && true === $options['with_error_data'] ) {
+				return new \WP_Error( $code, $message, $body );
 			}
 
 			return new \WP_Error( $code, $message );
@@ -574,6 +588,16 @@ abstract class Base_App {
 				'reconnect_nonce' => wp_create_nonce( $this->get_slug() . 'reconnect' ),
 			] );
 
+		$utm_campaign = get_transient( 'elementor_core_campaign' );
+
+		if ( ! empty( $utm_campaign ) ) {
+			foreach ( [ 'source', 'medium', 'campaign' ] as $key ) {
+				if ( ! empty( $utm_campaign[ $key ] ) ) {
+					$query_params->offsetSet( 'utm_' . $key, $utm_campaign[ $key ] );
+				}
+			}
+		}
+
 		return add_query_arg( $query_params->all(), $this->get_remote_site_url() );
 	}
 
@@ -592,6 +616,7 @@ abstract class Base_App {
 				break;
 
 			case 'cli':
+			case 'rest':
 				$this->admin_notice();
 				die;
 
@@ -650,6 +675,10 @@ abstract class Base_App {
 					<?php echo wp_json_encode( $data ); ?>
 				);
 
+				opener.dispatchEvent( new CustomEvent( 'elementor/connect/success' ),
+					<?php echo wp_json_encode( $data ); ?>
+				);
+
 				window.close();
 				opener.focus();
 			} else {
@@ -691,6 +720,7 @@ abstract class Base_App {
 	protected function redirect_to_remote_authorize_url() {
 		switch ( $this->auth_mode ) {
 			case 'cli':
+			case 'rest':
 				$this->get_app_token_from_cli_token( Utils::get_super_global_value( $_REQUEST, 'token' ) ); //phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Nonce verification is not required here.
 				return;
 			default:
@@ -722,6 +752,13 @@ abstract class Base_App {
 					printf( '[%s] %s', wp_kses_post( $notice['type'] ), wp_kses_post( $notice['content'] ) );
 				}
 				break;
+
+			case 'rest':
+				// After `wp_send_json` the script will die.
+				$this->delete( 'notices' );
+				wp_send_json( $notices );
+				break;
+
 			default:
 				/**
 				 * @var Admin_Notices $admin_notices
@@ -759,7 +796,6 @@ abstract class Base_App {
 			// PHPCS - the values of $item['label'], $color, $status are plain strings.
 			printf( '%s: <strong style="color:%s">%s</strong><br>', $item['label'], $color, $status ); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
 		}
-
 	}
 
 	private function get_generated_urls( $endpoint ) {
@@ -783,7 +819,7 @@ abstract class Base_App {
 			$this->set_auth_mode( 'xhr' );
 		}
 
-		$mode = Utils::get_super_global_value( $_REQUEST, 'mode' ); //phpcs:ignore WordPress.Security.NonceVerification.Recommended -- Nonce verification is not required here.
+		$mode = Utils::get_super_global_value( $_REQUEST, 'mode' );
 
 		if ( $mode ) {
 			$allowed_auth_modes = [
@@ -792,6 +828,10 @@ abstract class Base_App {
 
 			if ( defined( 'WP_CLI' ) && WP_CLI ) {
 				$allowed_auth_modes[] = 'cli';
+			}
+
+			if ( defined( 'REST_REQUEST' ) && REST_REQUEST ) {
+				$allowed_auth_modes[] = 'rest';
 			}
 
 			if ( in_array( $mode, $allowed_auth_modes, true ) ) {
